Create .gitignore and added .DS_Store file
The .DS_Store file is a metadata file created by macOS to store custom attributes of a folder. It is automatically created by Finder in each directory accessed and is hidden by default.
To prevent .DS_Store files from being tracked, it is good practice to add them to the .gitignore file.